+++ CHANGELOG.md
@@ -6,3 +6,4 @@ from VIVO
 * **2015-01-22** get is complete
 * **2015-01-27** General data structure for updates and updates.  Now handles datatype and lang.  Handles updates 
 and adds. Handles all six value conditions.
+* **2015-01-29** Enumerations handled.  Filters handled.